ejercicio3 v0.1.0 Ejercicio3

Documentation para el Ejercicio3.

Functions

get_bill(msg, n, i_inicio, i_final, res, divisor, ultimo_bueno, tope)

Parámtros n: número máximo de iteraciones, i_inicio: indice de inicio del mapa de 365 dias, i_final: número total de días, res: variable en donde se guarda la suma de facturas divisor: número de particiones, ultimo_bueno: último punto del mapa de donde se obtuvo resultados, tope: número de días del año. Función recursiva. Contempla las respuestas de la petición y en base a eso toma ciertas decisiones.

get_bills()

Obtener facturas : Obtiene facturas diviendo los 365 entre 2,3,4,.. hasta conseguir que sean menos 100 facturas. Parámtros n: número máximo de iteraciones, i_inicio: indice de inicio del mapa de 365 dias, i_final: número total de días, res: variable en donde se guarda la suma de facturas divisor: número de particiones, ultimo_bueno: último punto del mapa de donde se obtuvo resultados, tope: número de días del año.

Examples

iex> Ejercicio3.get_bills
:ejercicio3